+ this is a very far progress of work between commit. I've ommitted to commit the changes of Recite18th to master source.
I just summaries of what I've done lately.
1. FormGenerator.java
Able to create input of Primary Key field which is not autonumeric. The autonumeric will be hidden. And a pair hidden_**** will be created for primary key which is not autoincrement. Foreign key field will open a new window using ThichBox to choose from, so we need ChooserModel for this to work. Enumeration already displayed as SELECT field. But it may works only for MySQL database.

2. Model.java, Controller.java modify to add support for the editing of primary key which is not autoincrement